+++ source/uwp/UWPUITests/Properties/AssemblyInfo.cs
