+++ source/uwp/UWPUITests/UWPUITests.csproj
@@ -8,8 +8,8 @@
     <ProjectGuid>{C65B7488-ED39-4E9E-8A66-C53D72262660}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>UWPUITestLibrary</RootNamespace>
-    <AssemblyName>UWPUITestLibrary</AssemblyName>
+    <RootNamespace>UWPUITests</RootNamespace>
+    <AssemblyName>UWPUITests</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{3AC096D0-A1C2-E12C-1390-A8335801FDAB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
