+++ source/uwp/UWPUITests/packages.config
